requirements.txt
fastapi
uvicorn
faster-whisper
yt-dlp
numpy
sentence-transformers
summa
groq
python-dotenv
